Push from travis
Setup docker pushing by loggin in using ENVs that will be stored
encrypted in travis.

Find the true branch and only push if on master

Remove docker-compose testing

While it would be nice it currently breaks the whole build when
maven is down which stops the wdqs image from building.

We'd rather just one test fails