Accept a range of versions for apt and apk packages
In case the versions are available are increased upstream and the
version we were exactly pinned to become unavailable it is nice to
accept a sensible range.

With apt we use a wildcard and avoid going up more than a major
version.

With apk we can use the build in semver syntax with '~'