Relax pinned versions in quickstatements Dockerfile
Pinning exact versions means that the image fails to build if even the
smallest update to the package, since the old version might no longer be
available (for example, in Ubuntu Xenial the Git version recently
changed from 1:2.7.4-0ubuntu1.3 to 1:2.7.4-0ubuntu1.4). Instead, relax
the version requirements using globs, as done in other Dockerfiles.