wdqs: run as blazegraph without sudo
Run as blazegraph without resorting to sudo. According to what I
understood, a previous challenge was that /wdqs/data contained files
owned by root, but this should not be a problem if the volume gets
first created in a directory owned by the right user. Maybe this was
from a pre-existing volume?

Remark: the 0.3.6 Dockerfile missed the sudo installation so it could
not be removed there.

This is late follow-up to
https://github.com/wmde/wikibase-docker/pull/106